requirements.txt
aniso8601==9.0.1
attrs==22.2.0
click==8.1.3
coverage==7.2.1
exceptiongroup==1.1.0
Flask==2.2.3
Flask-RESTful==0.3.9
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
pluggy==1.0.0
pytest==7.2.2
pytest-cov==4.0.0
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
six==1.16.0
tomli==2.0.1
Werkzeug==2.2.3
zipp==3.15.0
